Refactor learning item cognitive load calculation and update schemas
- Moved the cognitive load calculation logic to a shared utility in `cognitive-load.ts` for better code reuse.
- Updated input schemas in `tools.ts` to utilize new schemas for learning items, session history, and session constraints, enhancing validation and maintainability.
- Simplified the recommendation engine's cognitive load checks by integrating the shared calculation function.
